Move settings store to settings
