+++ src/pages/graphData/GraphData.vue
@@ -77,7 +77,7 @@ import { ref, computed, onMounted, onUnmounted } from 'vue'
 import { emitter } from 'src/boot/mitt'
 import packageJson from '../../../package.json'
 import { useRefuelStore } from 'src/stores/refuelStore'
-import { useSettingsStore } from 'src/stores/settingsStore'
+import { useSettingsStore } from 'src/pages/settings/stores/settingsStore'
 import { useGraphDataStore } from './stores/graphDataStore'
 import { Period } from 'src/pages/graphData/scripts/models'
 import { optionsDialog } from 'src/components/dialogs/optionsDialog'
